CSCTTV-4277 TTV model update 2025_03. Move doi, isbn and issn from dim_publication to dim_pid.

+++ aspnetcore/src/DatabaseContext/Entities/FactFieldValuesTest.cs
@@ -9,7 +9,7 @@ public partial class FactFieldValuesTest
 
     public int DimFieldDisplaySettingsId { get; set; }
 
-    public long DimNameId { get; set; }
+    public int DimNameId { get; set; }
 
     public int DimWebLinkId { get; set; }
